End of Day 4/8/2018-home
Got mostly working 3.1.2. When coding the password validation I did not use String method charAt and get ASCII values of the characters in the password. Instead I used a collection of if statements and strings. May want to look into this and also switch statements. Also, the progress dialogue does not run at the moment. It is commented out. Look into this as well. :)